Edu applicant information page (#5065)
* WIP

* Pull out applicantInformation to pages

* Use direct deposit page again

* Modify applicantInformation()

* Capitalize the title

* Fix typo and use schema param

* Update unit test

* Use only schema from params
