+++ src/js/edu-benefits/5490/config/form.js
@@ -21,6 +21,7 @@ import { uiSchema as nonMilitaryJobsUiSchema } from '../../../common/schemaform/
 
 import createContactInformationPage from '../../pages/contactInformation';
 import directDeposit from '../../pages/directDeposit';
+import applicantInformation from '../../pages/applicantInformation';
 import createSchoolSelectionPage from '../../pages/schoolSelection';
 
 import IntroductionPage from '../components/IntroductionPage';
@@ -73,7 +74,14 @@ const formConfig = {
   chapters: {
     applicantInformation: {
       title: 'Applicant Information',
-      pages: {}
+      pages: {
+        applicantInformation: applicantInformation(fullSchema5490, [
+          'relativeFullName',
+          'relativeSocialSecurityNumber',
+          'relativeDateOfBirth',
+          'gender'
+        ])
+      }
     },
     benefitSelection: {
       title: 'Education Benefit',
