+++ test/edu-benefits/pages/applicantInformation.unit.spec.jsx
@@ -1,0 +1,129 @@
+import React from 'react';
+import { findDOMNode } from 'react-dom';
+import { expect } from 'chai';
+import sinon from 'sinon';
+import ReactTestUtils from 'react-addons-test-utils';
+
+import { DefinitionTester, submitForm } from '../../util/schemaform-utils.jsx';
+import formConfig from '../../../src/js/edu-benefits/pages/applicantInformation.js';
+
+import fullSchema1990e from 'vets-json-schema/dist/transfer-benefits-schema.json';
+import fullSchema5490 from 'vets-json-schema/dist/dependents-benefits-schema.json';
+
+describe('Edu 1990e applicantInformation', () => {
+  const { schema, uiSchema } = formConfig(fullSchema1990e);
+  it('should render', () => {
+    const form = ReactTestUtils.renderIntoDocument(
+      <DefinitionTester
+          schema={schema}
+          data={{}}
+          uiSchema={uiSchema}/>
+    );
+    const formDOM = findDOMNode(form);
+
+    expect(formDOM.querySelectorAll('input, select').length).to.equal(12);
+  });
+  it('should show errors when required fields are empty', () => {
+    const onSubmit = sinon.spy();
+    const form = ReactTestUtils.renderIntoDocument(
+      <DefinitionTester
+          schema={schema}
+          onSubmit={onSubmit}
+          data={{}}
+          uiSchema={uiSchema}/>
+    );
+    const formDOM = findDOMNode(form);
+    submitForm(form);
+    expect(Array.from(formDOM.querySelectorAll('.usa-input-error'))).not.to.be.empty;
+    expect(onSubmit.called).not.to.be.true;
+  });
+  it('should show no errors when all required fields are filled', () => {
+    const onSubmit = sinon.spy();
+    const form = ReactTestUtils.renderIntoDocument(
+      <DefinitionTester
+          schema={schema}
+          onSubmit={onSubmit}
+          data={{}}
+          uiSchema={uiSchema}/>
+    );
+
+    const formDOM = findDOMNode(form);
+    const find = formDOM.querySelector.bind(formDOM);
+    ReactTestUtils.Simulate.change(find('#root_relativeFullName_first'), {
+      target: {
+        value: 'Test'
+      }
+    });
+    ReactTestUtils.Simulate.change(find('#root_relativeFullName_last'), {
+      target: {
+        value: 'Test'
+      }
+    });
+
+    submitForm(form);
+    expect(Array.from(formDOM.querySelectorAll('.usa-input-error'))).to.be.empty;
+    expect(onSubmit.called).to.be.true;
+  });
+});
+
+describe('Edu 5490 applicantInformation', () => {
+  // Make sure these match the fields listed in the 5490
+  const { schema, uiSchema } = formConfig(fullSchema5490, [
+    'relativeFullName',
+    'relativeSocialSecurityNumber',
+    'relativeDateOfBirth',
+    'gender'
+  ]);
+  it('should render', () => {
+    const form = ReactTestUtils.renderIntoDocument(
+      <DefinitionTester
+          schema={schema}
+          data={{}}
+          uiSchema={uiSchema}/>
+    );
+    const formDOM = findDOMNode(form);
+
+    expect(formDOM.querySelectorAll('input, select').length).to.equal(10);
+  });
+  it('should show errors when required fields are empty', () => {
+    const onSubmit = sinon.spy();
+    const form = ReactTestUtils.renderIntoDocument(
+      <DefinitionTester
+          schema={schema}
+          onSubmit={onSubmit}
+          data={{}}
+          uiSchema={uiSchema}/>
+    );
+    const formDOM = findDOMNode(form);
+    submitForm(form);
+    expect(Array.from(formDOM.querySelectorAll('.usa-input-error'))).not.to.be.empty;
+    expect(onSubmit.called).not.to.be.true;
+  });
+  it('should show no errors when all required fields are filled', () => {
+    const onSubmit = sinon.spy();
+    const form = ReactTestUtils.renderIntoDocument(
+      <DefinitionTester
+          schema={schema}
+          onSubmit={onSubmit}
+          data={{}}
+          uiSchema={uiSchema}/>
+    );
+
+    const formDOM = findDOMNode(form);
+    const find = formDOM.querySelector.bind(formDOM);
+    ReactTestUtils.Simulate.change(find('#root_relativeFullName_first'), {
+      target: {
+        value: 'Test'
+      }
+    });
+    ReactTestUtils.Simulate.change(find('#root_relativeFullName_last'), {
+      target: {
+        value: 'Test'
+      }
+    });
+
+    submitForm(form);
+    expect(Array.from(formDOM.querySelectorAll('.usa-input-error'))).to.be.empty;
+    expect(onSubmit.called).to.be.true;
+  });
+});
